requirements.txt
mcp
fastmcp
transformers 
torchaudio 
librosa 
soundfile
deepface
tf-keras
speechbrain 
torchaudio